gemini: convert to TS
